Refactor Thumbnails
