+++ client/src/components/Overview/index.jsx
@@ -18,7 +18,13 @@ function Overview() {
   const [activeStyleId] = useContext(ActiveStyleId);
   const activeStyle = getStyle(currentStyles, activeStyleId);
 
-  const [galleryView] = useState('default');
+  const [galleryView, setGalleryView] = useState('default');
+
+  const handleExpandedView = (e) => {
+    if (e.target === e.currentTarget && galleryView === 'default') {
+      setGalleryView('expanded');
+    }
+  };
 
   // Only render the Overview component if there's an active style.
   // Eventually, add more graceful loading.
@@ -28,8 +34,13 @@ function Overview() {
       <PreviewStyleProvider>
         <ExpandedImageGallery>
 
-          <DefaultImageGallery view={galleryView}>
-            <ImageGallery view={galleryView} />
+          <DefaultImageGallery>
+            <DefaultImageGalleryViewport className={galleryView}>
+              <ImageGallery
+                view={galleryView}
+                handleExpandedView={handleExpandedView}
+              />
+            </DefaultImageGalleryViewport>
           </DefaultImageGallery>
 
           <ProductInfo>
@@ -49,17 +60,28 @@ function Overview() {
 const ExpandedImageGallery = styled.section`
   background-color: ${(props) => props.theme.colors.light};
   display: flex;
-  min-height: 630px;
+  height: 630px;
 `;
 
 const DefaultImageGallery = styled.div`
+  width: 800px;
+  overflow: visible;
+  z-index: 2;
+`;
+
+const DefaultImageGalleryViewport = styled.div`
   background-color:${(props) => props.theme.colors.background};
-  width: 67.5%;
+  width: 100%;
+  height: 100%;
+  &.expanded {
+    width: 1280px;
+    transition: width 1s ease-in-out;
+  }
 `;
 
 const ProductInfo = styled.section`
   background-color:${(props) => props.theme.colors.light};
-  width: 32.5%;
+  max-width: 480px;
   padding: 10px;
 `;
